==> 3_building_RF_for_HALC.ipynb <==
In this notebook, we use clean_data.csv to create a model to predict HALC
- RUN ORDER = 3/5

import pandas as pd
import numpy as np
from sklearn.metrics import roc_curve, roc_auc_score
from sklearn.linear_model import LogisticRegression
import warnings
warnings.filterwarnings('ignore')
from sklearn.metrics import classification_report, confusion_matrix
from imblearn.over_sampling import SMOTE
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_squared_error

df = pd.read_csv("clean_data.csv")

- We will create a RF to predict HALC using all columns except made_claim

- Again, we are parsing data where HALC is not 0
  - Is it heavily unbalanced
  - We want predicted HALC to a continuous number because the 0's would have already been predicted

df = df[df['HALC'] != 0]

df = df.drop(columns=['made_claim'])

# remove outliers in df to better represent the data
Q1 = df['LC'].quantile(0.25)
Q3 = df['LC'].quantile(0.75)
IQR = Q3 - Q1

lower_bound = Q1 - 1.5 * IQR
upper_bound = Q3 + 1.5 * IQR

df = df[(df['LC'] >= lower_bound) & (df['LC'] <= upper_bound)]

# temp = df.pop('LC')
# df['LC'] = temp

df.columns

X = df.drop(columns=['HALC'])
y = df['HALC']

X_train, X_test, y_train, y_test = train_test_split(
    X, y, test_size=0.2, random_state=42
)

rf_HALC = RandomForestRegressor(n_estimators=100, random_state=42)
rf_HALC.fit(X_train, y_train)

y_pred = rf_HALC.predict(X_test)

- RMSE: 369.99

mean_squared_error(y_test, y_pred, squared=False)

import joblib
joblib.dump(rf_HALC, 'predict_HALC.pkl')

X_test.columns

y

df

np.mean(df['LC'])
